feat(ux): reorganizar sidebar Riesgos + extraer Calculadora USDCOP
PR 1 + 2 consolidados del trabajo de UX/IA del modulo de Riesgos.

## PR 1 — Sidebar reorganizado (workflow-oriented)

Antes: 9 entradas planas en grupo Riesgos sin agrupacion logica.
Ahora: 4 secciones visualmente separadas por dividers:

  Riesgos
  ├── Resumen
  │   POSICIONES
  ├── Commodities (era 'Exposicion')
  ├── Portafolio OTC
  ├── Creditos
  ├── Portafolio TES
  │   CALCULADORAS
  ├── NDF Pricer
  ├── IBR Swap
  ├── XCCY Swap
  ├── COLTES
  ├── USDCOP                    ← NUEVA
  │   DATOS DE MERCADO
  └── Marcas                    ← NUEVA

Cambios:
- src/layout/CoreLayout/Sidebar/SidebarNavList.tsx:
  - RIESGOS_SUBNAV tipado como `NavItemProps | { divider, name }`
  - Render loop maneja dividers como section headers (no navegables)
  - Renombrado: Exposicion -> Commodities, anadido USDCOP + Marcas
  - active path detection ampliado con /usdcop-calculator y /marks
- src/layout/CoreLayout/CoreLayout.tsx: RISK_PATHS incluye las 2 rutas
  nuevas para que el GlobalDateSelector aparezca en ellas.

## PR 2 — Calculadora USDCOP extraida a /usdcop-calculator

Antes: vivia como ultimo tab dentro de /risk-management. El usuario
nuevo no la encontraba ('por que esta en Exposicion una calculadora de
FX?').

Ahora: pagina standalone propia. Misma jerarquia que NDF / IBR / XCCY
/ COLTES Pricers.

Cambios:
- src/pages/usdcop-calculator/index.tsx (NUEVO): UI completa de la
  calculadora con CoreLayout + RoleGuard + PageTitle. Ancla al
  selector global de fecha igual que antes.
- src/pages/risk-management/index.tsx: removido
  * tab 'usdcop' del baseTabs builder
  * state usdcopData/Loading/Days/Sigma
  * handler handleFetchUsdcop
  * dispatch en useEffect del activeTab
  * JSX block del tab (~250 lineas)
  * imports huerfanos: faCalculator, ComposedChart, Area,
    fetchUsdCopCalculator, UsdCopData

## Companion futuro
- PR 3 (proximo): Marcas — quitar el tab de /portfolio (la entrada
  /marks ya esta en sidebar).
